finalizado crud da unidade e atendimento / melhorias gerais

+++ healthCare-backend/src/main/java/br/spin/crud/unidades/repository/UnidadeRepository.java
@@ -3,5 +3,5 @@ package br.spin.crud.unidades.repository;
 import br.spin.crud.unidades.models.Unidade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnidadeRepository extends JpaRepository<Unidade, Integer> {
+public interface UnidadeRepository extends JpaRepository<Unidade, Long> {
 }
